@@ -59,14 +59,18 @@
     }
   },
   "resolutions": {
-    "@types/react": "^18.2.36",
-    "react-is": "^18.2.0",
     "@babel/runtime": "^7",
+    "@typescript-eslint/eslint-plugin": "^6",
+    "@typescript-eslint/parser": "^6",
+    "@typescript-eslint/scope-manager": "^6",
+    "@typescript-eslint/type-utils": "^6",
+    "@typescript-eslint/types": "^6",
+    "@typescript-eslint/typescript-": "^6",
+    "@typescript-eslint/utils": "^6",
+    "@typescript-eslint/visitor-keys": "^6",
+    "clsx": "^2.0.0",
     "nth-check": "^2.0.1",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
-    "fork-ts-checker-webpack-plugin": "^6.5.3",
-    "rxjs": "^7.8.1"
+    "react-is": "^18.2.0"
   },
   "packageManager": "yarn@4.0.2",
   "scripts": {
@@ -88,16 +92,16 @@
     "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-replace": "^5.0.5",
-    "@types/node": "^20.10.4",
+    "@types/node": "^20.10.8",
     "@types/rimraf": "^4.0.5",
     "build-if-changed": "^1.5.5",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-prettier": "^5.1.3",
     "glob-gitignore": "^1.0.14",
     "npm-run-all": "^4.1.5",
     "rimraf": "^5.0.5",
-    "rollup": "^4.7.0",
+    "rollup": "^4.9.4",
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-terser": "^7.0.2",
