@@ -36,10 +36,10 @@
   "devDependencies": {
     "@babel/cli": "^7.23.4",
     "@babel/plugin-transform-object-assign": "^7.23.3",
-    "@types/eslint": "^8.44.8",
+    "@types/eslint": "^8.56.1",
     "babel-plugin-inline-package-json": "^2.0.0",
     "build-if-changed": "^1.5.5",
-    "eslint": "^8.55.0"
+    "eslint": "^8.56.0"
   },
   "peerDependencies": {
     "uppy": "^2.13.8"
