@@ -15,13 +15,20 @@
  */
 
 import * as React from 'react';
-import EditModesSwitcherUI from './EditModesSwitcherUI';
 import { useActiveUser } from '../../hooks/useActiveUser';
 import { isItemLockedForMe } from '../../utils/content';
 import { useDispatch } from 'react-redux';
 import { setPreviewEditMode } from '../../state/actions/preview';
 import { DetailedItem } from '../../models/Item';
 import { usePreviewState } from '../../hooks/usePreviewState';
+import Box from '@mui/material/Box';
+import Tooltip from '@mui/material/Tooltip';
+import { FormattedMessage } from 'react-intl';
+import IconButton from '@mui/material/IconButton';
+import { UNDEFINED } from '../../utils/constants';
+import PowerSettingsNewRoundedIcon from '@mui/icons-material/PowerSettingsNewRounded';
+import EditRoundedIcon from '@mui/icons-material/EditRounded';
+import DragIndicatorRoundedIcon from '@mui/icons-material/DragIndicatorRounded';
 
 export interface EditModesSwitcherProps {
   item: DetailedItem;
@@ -35,15 +42,98 @@ export function EditModesSwitcher(props: EditModesSwitcherProps) {
   const write = Boolean(item?.availableActionsMap.edit);
   const dispatch = useDispatch();
   const { editMode, highlightMode } = usePreviewState();
-  const onChange = (editMode, highlightMode) => dispatch(setPreviewEditMode({ editMode, highlightMode }));
-  const isDisabled = disabled || !write || isLocked;
+  const onEditModeChange = (editMode, highlightMode?) => dispatch(setPreviewEditMode({ editMode, highlightMode }));
+  const isComputedDisabled = disabled || !write || isLocked;
+  const isAllHighlightMode = editMode && highlightMode === 'all';
+  const isMoveHighlightMode = editMode && !isAllHighlightMode;
+  const getStyle = () => ({
+    bgcolor: 'background.default',
+    ':hover': {
+      cursor: 'default'
+    }
+  });
   return (
-    <EditModesSwitcherUI
-      disabled={isDisabled}
-      isEditMode={editMode}
-      highlightMode={highlightMode}
-      onEditModeChange={onChange}
-    />
+    <Box
+      sx={{
+        minWidth: 104,
+        borderRadius: 20,
+        display: 'inline-block',
+        border: (theme) =>
+          `1px solid ${theme.palette.mode === 'light' ? theme.palette.divider : theme.palette.grey[700]}`,
+        transition: (theme) => theme.transitions.create(['background-color', 'border'])
+      }}
+    >
+      <Tooltip
+        title={
+          <FormattedMessage
+            id="editModesSwitcher.offButtonTooltip"
+            defaultMessage="Switch off editing ({shortcutKey})"
+            values={{
+              shortcutKey: editMode ? (isAllHighlightMode ? 'e' : 'm') : 'e | m'
+            }}
+          />
+        }
+      >
+        <IconButton
+          size="small"
+          color={!editMode ? 'error' : UNDEFINED}
+          disabled={disabled}
+          onClick={() => onEditModeChange(false)}
+          sx={{
+            borderTopRightRadius: 0,
+            borderBottomRightRadius: 0,
+            ...(!editMode && { cursor: 'default' })
+          }}
+        >
+          <PowerSettingsNewRoundedIcon />
+        </IconButton>
+      </Tooltip>
+      <Tooltip
+        title={
+          isComputedDisabled ? (
+            ''
+          ) : (
+            <FormattedMessage id="editModesSwitcher.editModeTooltip" defaultMessage="Edit mode (e)" />
+          )
+        }
+      >
+        <IconButton
+          color={isAllHighlightMode ? 'success' : UNDEFINED}
+          disabled={isComputedDisabled}
+          size="small"
+          onClick={() => onEditModeChange(true, 'all')}
+          sx={{
+            borderRadius: 0,
+            ...(isAllHighlightMode && getStyle())
+          }}
+        >
+          <EditRoundedIcon />
+        </IconButton>
+      </Tooltip>
+      <Tooltip
+        title={
+          isComputedDisabled ? (
+            ''
+          ) : (
+            <FormattedMessage id="editModesSwitcher.moveModeTooltip" defaultMessage="Move mode (m)" />
+          )
+        }
+      >
+        <IconButton
+          color={isMoveHighlightMode ? 'primary' : UNDEFINED}
+          disabled={isComputedDisabled}
+          size="small"
+          onClick={() => onEditModeChange(true, 'move')}
+          sx={{
+            borderTopLeftRadius: 0,
+            borderBottomLeftRadius: 0,
+            ...(isMoveHighlightMode && getStyle())
+          }}
+        >
+          <DragIndicatorRoundedIcon />
+        </IconButton>
+      </Tooltip>
+    </Box>
   );
 }
 
