@@ -19,17 +19,12 @@ import { useCallback, useEffect, useRef } from 'react';
 import queryString from 'query-string';
 import { LookupTable } from '../models/LookupTable';
 import { changeCurrentUrl } from '../state/actions/preview';
-import { changeSite, popSite } from '../state/actions/sites';
+import { changeSite } from '../state/actions/sites';
 import { useActiveSiteId } from './useActiveSiteId';
 import { useEnv } from './useEnv';
 import { usePreviewNavigation } from './usePreviewNavigation';
 import useSiteLookup from './useSiteLookup';
 import { defineMessages, useIntl } from 'react-intl';
-import { getHostToHostBus } from '../utils/subjects';
-import { filter } from 'rxjs/operators';
-import { projectDeleted } from '../state/actions/system';
-import { ProjectLifecycleEvent } from '../models/ProjectLifecycleEvent';
-import StandardAction from '../models/StandardAction';
 
 const messages = defineMessages({
   siteNotFound: {
@@ -37,9 +32,6 @@ const messages = defineMessages({
   },
   siteNotReady: {
     defaultMessage: 'Project not initialized yet. Redirecting to projects list.'
-  },
-  siteDeleted: {
-    defaultMessage: "This project has been deleted, you'll be redirected to projects list."
   }
 });
 
@@ -86,21 +78,6 @@ export function usePreviewUrlControl(history) {
     },
     [sites, authoringBase, formatMessage]
   );
-
-  useEffect(() => {
-    const hostToHost$ = getHostToHostBus();
-    const subscription = hostToHost$
-      .pipe(filter((e) => e.type === projectDeleted.type))
-      .subscribe(({ payload }: StandardAction<ProjectLifecycleEvent<'SITE_DELETED_EVENT'>>) => {
-        if (payload.siteUuid === sites[site]?.uuid) {
-          notValidSiteRedirect(formatMessage(messages.siteDeleted), `${authoringBase}#/sites`);
-          dispatch(popSite({ siteId: site }));
-        }
-      });
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, [authoringBase, formatMessage, site, dispatch, sites]);
 
   useEffect(() => {
     const prev = priorState.current;
