@@ -336,7 +336,7 @@ function performMutation(
           post(
             writeContentUrl({
               site,
-              path: path,
+              path,
               unlock: 'true',
               fileName: getInnerHtml(doc.querySelector(':scope > file-name'))
             }),
@@ -981,7 +981,11 @@ export function createFileUpload(
   metaData: Record<string, unknown>,
   xsrfArgumentName: string
 ): Observable<StandardAction> {
-  const qs = toQueryString({ [xsrfArgumentName]: getRequestForgeryToken() });
+  const qs = toQueryString({
+    path,
+    site: metaData?.site ?? metaData?.siteId,
+    [xsrfArgumentName]: getRequestForgeryToken()
+  });
   return new Observable((subscriber) => {
     const uppy = new Core({ autoProceed: true });
     uppy.use(XHRUpload, { endpoint: `${uploadUrl}${qs}`, headers: getGlobalHeaders() });
@@ -1008,8 +1012,19 @@ export function createFileUpload(
       });
     });
 
-    uppy.on('upload-error', (file, error) => {
-      subscriber.error(error);
+    uppy.on('upload-error', (file, error, response) => {
+      // @ts-ignore
+      response.error = response;
+      subscriber.error(
+        response
+        // type CustomUploadError {
+        //   error: { request: XMLHttpRequest } & Error;
+        //   body: {
+        //     response: ApiResponse;
+        //     status: number;
+        //   };
+        // }
+      );
     });
 
     uppy.addFile({
