@@ -15,7 +15,13 @@ import {
 import { AjaxError } from 'rxjs/ajax';
 import { SHARED_WORKER_NAME, XSRF_TOKEN_HEADER_NAME } from './utils/constants';
 import { Client, StompSubscription } from '@stomp/stompjs';
-import { emitSystemEvent, globalSocketStatus, openSiteSocket, siteSocketStatus } from './state/actions/system';
+import {
+  closeSiteSocket,
+  emitSystemEvent,
+  globalSocketStatus,
+  openSiteSocket,
+  siteSocketStatus
+} from './state/actions/system';
 
 declare const self: SharedWorkerGlobalScope;
 
@@ -94,6 +100,9 @@ function onmessage(event) {
     case openSiteSocket.type:
       openSocket(event.data.payload);
       break;
+    case closeSiteSocket.type:
+      closeSocket(event.data.payload.site);
+      break;
     default:
       log(`Received unknown action: "${type}"`);
       break;
@@ -169,20 +178,22 @@ function broadcast(message: StandardAction, excludedClient?: MessagePort) {
   });
 }
 
+function broadcastSocketConnection(connected: boolean, siteId: string = null) {
+  broadcast(siteId ? siteSocketStatus({ siteId, connected }) : globalSocketStatus({ connected }));
+}
+
 function openSocket({ site, xsrfToken }) {
   let isSiteSocket = !!site;
   let socketClient = isSiteSocket ? siteSocketClient : rootSocketClient;
   socketClient?.deactivate();
   let subscription: StompSubscription;
   let protocol = self.location.protocol === 'https:' ? 'wss' : 'ws';
-  let broadcastConnection = (connected: boolean) =>
-    broadcast(isSiteSocket ? siteSocketStatus({ siteId: site, connected }) : globalSocketStatus({ connected }));
   socketClient = new Client({
     brokerURL: `${protocol}://${isProduction ? self.location.host : 'localhost:8080'}/studio/events`,
     ...(!isProduction && { debug: log }),
     connectHeaders: { [XSRF_TOKEN_HEADER_NAME]: xsrfToken },
     onConnect() {
-      broadcastConnection(true);
+      broadcastSocketConnection(true, site);
       const topicUrl = isSiteSocket ? `/topic/studio/${site}` : '/topic/studio';
       subscription = socketClient.subscribe(topicUrl, (message) => {
         if (message.body) {
@@ -203,10 +214,10 @@ function openSocket({ site, xsrfToken }) {
     onStompError() {
       // Will be invoked in case of error encountered at Broker
       // Bad login/passcode typically will cause an error
-      broadcastConnection(false);
+      broadcastSocketConnection(false, site);
     },
     onWebSocketError() {
-      broadcastConnection(false);
+      broadcastSocketConnection(false, site);
     },
     onDisconnect() {
       subscription?.unsubscribe();
@@ -218,6 +229,18 @@ function openSocket({ site, xsrfToken }) {
     siteSocketClient = socketClient;
   } else {
     rootSocketClient = socketClient;
+  }
+}
+
+function closeSocket(siteId?: string) {
+  if (siteId) {
+    siteSocketClient?.deactivate();
+    siteSocketClient = null;
+    broadcastSocketConnection(false, siteId);
+  } else {
+    rootSocketClient?.deactivate();
+    rootSocketClient = null;
+    broadcastSocketConnection(false);
   }
 }
 
