@@ -42,6 +42,7 @@ import {
   initRichTextEditorConfig,
   initToolbarConfig,
   initToolsPanelConfig,
+  mainModelModifiedExternally,
   openToolsPanel,
   popIcePanelPage,
   popToolsPanelPage,
@@ -165,12 +166,11 @@ const initialState: GlobalState['preview'] = {
   richTextEditor: null,
   editModePadding: false,
   windowSize: window.innerWidth,
-  xbDetectionTimeoutMs: null
+  xbDetectionTimeoutMs: 5000
 };
 
 const minDrawerWidth = 240;
 const minPreviewWidth = 320;
-const defaultXbDetectionTimeoutMs = 5000;
 
 const isDrawerWidthValid = (
   windowSize: number,
@@ -253,7 +253,6 @@ const reducer = createReducer<GlobalState['preview']>(initialState, (builder) =>
       const previewConfigEl = configDOM.querySelector('[id="craftercms.components.Preview"]');
       const initialEditModeOn = previewConfigEl?.getAttribute('initialEditModeOn');
       const initialHighlightMode = previewConfigEl?.getAttribute('initialHighlightMode') as HighlightMode;
-      const xbDetectionTimeoutMs = parseInt(previewConfigEl?.getAttribute('xbDetectionTimeoutMs'));
 
       // If there is no storedEditMode, set it to the value of initialEditModeOn (config value), otherwise, defaults to true
       state.editMode = payload.storedEditMode ?? (initialEditModeOn ? initialEditModeOn === 'true' : true);
@@ -261,10 +260,6 @@ const reducer = createReducer<GlobalState['preview']>(initialState, (builder) =>
         payload.storedHighlightMode ??
         (['all', 'move'].includes(initialHighlightMode) ? initialHighlightMode : state.highlightMode);
       state.editModePadding = payload.storedPaddingMode ?? state.editModePadding;
-      state.xbDetectionTimeoutMs =
-        !isNaN(xbDetectionTimeoutMs) && xbDetectionTimeoutMs >= 0
-          ? xbDetectionTimeoutMs
-          : state.xbDetectionTimeoutMs ?? defaultXbDetectionTimeoutMs;
     })
     .addCase(openToolsPanel, (state) => {
       const { windowSize, editMode, toolsPanelWidth, icePanelWidth } = state;
@@ -327,7 +322,8 @@ const reducer = createReducer<GlobalState['preview']>(initialState, (builder) =>
         hierarchyMap: null,
         modelIdByPath: null,
         selected: null,
-        itemBeingDragged: null
+        itemBeingDragged: null,
+        mainModelModifier: null
       };
     })
     .addCase(guestCheckOut, (state) => {
@@ -800,6 +796,9 @@ const reducer = createReducer<GlobalState['preview']>(initialState, (builder) =>
       state.windowSize = windowSize;
       state.toolsPanelWidth = adjustedToolsPanelWidth;
       state.icePanelWidth = adjustedIcePanelWidth;
+    })
+    .addCase(mainModelModifiedExternally, (state, { payload }) => {
+      if (state.guest) state.guest.mainModelModifier = payload.user;
     });
 });
 
