@@ -42,7 +42,9 @@ export function initTinyMCE(
   const type = field?.type;
   const inlineElsRegex =
     /^(B|BIG|I|SMALL|TT|ABBR|ACRINYM|CITE|CODE|DFN|EM|KBD|STRONG|SAMP|VAR|A|BDO|BR|IMG|MAP|OBJECT|Q|SCRIPT|SPAN|SUB|SUP|BUTTON|INPUT|LABEL|SELECT|TEXTAREA)$/;
-  let rteEl = record.element;
+  const originalElement = record.element;
+  const originalRawContent = originalElement.innerHTML;
+  let rteEl = originalElement;
   const isRecordElInline = record.element.tagName.match(inlineElsRegex);
 
   // If record element is of type inline (doesn't matter the display prop), replace it with a block element (div).
@@ -172,23 +174,39 @@ export function initTinyMCE(
     },
     setup(editor: Editor) {
       let changed = false;
-      let originalContent;
-      let pluginManager = window.tinymce.util.Tools.resolve('tinymce.PluginManager');
+      const pluginManager = window.tinymce.util.Tools.resolve('tinymce.PluginManager');
+      const nonChars = [
+        'Meta',
+        'Alt',
+        'Control',
+        'Shift',
+        'CapsLock',
+        'Tab',
+        'Escape',
+        'ArrowLeft',
+        'ArrowRight',
+        'ArrowUp',
+        'ArrowDown',
+        'Dead',
+        'Delete'
+        // Added as needed when using this array...
+        // 'Backspace',
+        // 'Enter'
+      ].filter(Boolean);
 
       function save() {
         const content = getContent();
-        // 'change' event is not triggering until focusing out in v6. Reported in here https://github.com/tinymce/tinymce/issues/9132
-        if (changed || content !== originalContent) {
+        if (changed) {
           contentController.updateField(record.modelId, record.fieldId[0], record.index, content);
         }
       }
 
       function getContent() {
-        return type === 'html' ? editor.getContent() : editor.getContent({ format: 'text' });
+        return editor.getContent({ format: type === 'html' ? 'html' : 'text' });
       }
 
       function getSelectionContent() {
-        return type === 'html' ? editor.selection.getContent() : editor.selection.getContent({ format: 'text' });
+        return editor.selection.getContent({ format: type === 'html' ? 'html' : 'text' });
       }
 
       function destroyEditor() {
@@ -196,22 +214,19 @@ export function initTinyMCE(
       }
 
       function cancel({ saved }: { saved: boolean }) {
-        const content = getContent();
+        const finalContent = saved ? getContent() : originalRawContent;
+
         destroyEditor();
 
-        // In case the user did some text bolding or other formatting which won't
-        // be honoured on plain text, revert the content to the edited plain text
-        // version of the input.
-        changed && type === 'text' && (record.element.innerHTML = content);
+        originalElement.innerHTML = finalContent;
 
         if (isRecordElInline) {
-          // Update original element and remove created blockElement
-          record.element.innerHTML = rteEl.innerHTML;
+          // Remove the created blockElement and remove the display: none on original element
           rteEl.remove();
           record.element.style.display = '';
         }
 
-        if (record.element.innerHTML.trim() === '') {
+        if (finalContent.trim() === '') {
           record.element.classList.add(emptyFieldClass);
         }
 
@@ -224,8 +239,16 @@ export function initTinyMCE(
         }, 150);
       }
 
+      function replaceLineBreaksIfApplicable(content: string) {
+        // Replace line breaks with <br> for textarea fields
+        // Address line breaks in textarea fields: https://github.com/craftercms/craftercms/issues/6432
+        type === 'textarea' && editor.setContent(content.replaceAll('\n', '<br>'), { format: 'html' });
+      }
+
       editor.on('init', function () {
-        originalContent = getContent();
+        const initialTinyContent = getContent();
+
+        replaceLineBreaksIfApplicable(originalRawContent);
 
         editor.focus(false);
         editor.selection.select(editor.getBody(), true);
@@ -236,6 +259,8 @@ export function initTinyMCE(
         editor.on('focusout', (e) => {
           let saved = false;
           let relatedTarget = e.relatedTarget as HTMLElement;
+          // The 'change' event is not triggering until focusing out in v6. Reported in here https://github.com/tinymce/tinymce/issues/9132
+          changed = changed || getContent() !== initialTinyContent;
           if (
             !relatedTarget?.closest('.tox-tinymce') &&
             !relatedTarget?.classList.contains('tox-dialog__body-nav-item')
@@ -248,16 +273,9 @@ export function initTinyMCE(
                   values: { field: record.label }
                 })
               );
-              editor.setContent(originalContent);
-            } else {
-              // Replace new line char which is not honoured if the field is not html type
-              if (type !== 'html') {
-                editor.setContent(getContent().replace(/\n+/g, ' '));
-              }
-              if (getContent() !== originalContent) {
-                saved = true;
-                save();
-              }
+            } else if (changed) {
+              saved = true;
+              save();
             }
             e.stopImmediatePropagation();
             cancel({ saved });
@@ -282,37 +300,74 @@ export function initTinyMCE(
         }
       });
 
-      // TODO: The max-length can be bypassed by pasting.
-      // editor.on('paste', (e) => {
-      //   // ...
-      // });
+      editor.on('paste', (e) => {
+        const maxLength = validations?.maxLength ? parseInt(validations.maxLength.value) : null;
+        const text = (
+          e.clipboardData ||
+          // @ts-ignore
+          window.clipboardData
+        ).getData('text');
+        console.log(`"${text}"`, text.length, maxLength);
+        if (maxLength && text.length > maxLength) {
+          post(
+            snackGuestMessage({
+              id: 'maxLength',
+              level: 'required',
+              values: { maxLength: text.length === maxLength ? text.length : `${text.length}/${maxLength}` }
+            })
+          );
+        }
+        if (type === 'textarea') {
+          // Doing this immediately (without the timeout) causes the content to be duplicated.
+          // TinyMCE seems to be doing something internally that causes this.
+          setTimeout(() => {
+            replaceLineBreaksIfApplicable(text);
+            editor.selection.select(editor.getBody(), true);
+            editor.selection.collapse(false);
+          }, 10);
+        }
+        // TODO: It'd be great to be able to select the piece of the pasted content that falls out of the max-length.
+      });
+
+      editor.on('keyup', (e) => {
+        let content = getContent();
+        if (validations?.required && content.trim() === '' && !nonChars.concat('Enter').includes(e.key)) {
+          post(
+            snackGuestMessage({
+              id: 'required',
+              level: 'suggestion',
+              values: { field: record.label }
+            })
+          );
+        }
+      });
 
       editor.on('keydown', (e) => {
+        let content: string, selection: string, numMaxLength: number;
         if (e.key === 'Escape') {
           e.stopImmediatePropagation();
-          editor.setContent(originalContent);
           cancel({ saved: false });
         } else if ((e.ctrlKey || e.metaKey) && e.key === 'Enter') {
           e.preventDefault();
           // Timeout to avoid "Uncaught TypeError: Cannot read properties of null (reading 'getStart')"
           // Hypothesis is the focusout destroys the editor before some internal tiny thing runs.
           setTimeout(() => editor.fire('focusout'));
-        } else if (e.key === 'Enter' && type !== 'html') {
+        } else if (e.key === 'Enter' && type !== 'html' && type !== 'textarea') {
+          // Avoid new line in plain text fields
           e.preventDefault();
         } else if (
           validations?.maxLength &&
-          // TODO: Check/improve regex
-          /[a-zA-Z0-9-_ ]/.test(String.fromCharCode(e.keyCode)) &&
-          getContent().length + 1 > parseInt(validations.maxLength.value) &&
+          !nonChars.concat('Backspace').includes(e.key) &&
+          (content = getContent()).length + 1 > (numMaxLength = parseInt(validations.maxLength.value)) &&
           // If everything is selected and a key is pressed, essentially, it will
           // delete everything so no max-length problem
-          getContent() !== getSelectionContent()
+          ((selection = getSelectionContent()) === '' || content.length - (selection.length + 1) > numMaxLength)
         ) {
           post(
             snackGuestMessage({
               id: 'maxLength',
               level: 'required',
-              values: { maxLength: validations.maxLength.value }
+              values: { maxLength: `${content.length}/${validations.maxLength.value}` }
             })
           );
           e.stopPropagation();
